% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointLTSM.R
\name{jointLTSM}
\alias{jointLTSM}
\title{Estimation of latent time shift model for multivariate longitudinal 
outcomes (and time-to-event data) with a time-shifting process.}
\usage{
jointLTSM(
  fixed,
  random,
  range,
  link,
  intnodes = NULL,
  model,
  timevar,
  fixed.s,
  subject,
  data,
  subset = NULL,
  na.action = 1,
  B,
  posfix = NULL,
  survival = NULL,
  hazard = "Weibull",
  hazardrange = NULL,
  hazardnodes = NULL,
  logscale = FALSE,
  startWeibull = 0,
  surv.shift = NULL,
  sharedtype = "RE",
  nGK = 15,
  methInteg = "QMC",
  nMC = 1000,
  maxiter = 100,
  convB = 1e-04,
  convL = 1e-04,
  convG = 1e-04,
  partialH = NULL,
  nsim = 100,
  verbose = TRUE,
  returndata = FALSE,
  nproc = 1,
  clustertype = NULL
)
}
\arguments{
\item{fixed}{a list containing two-sided linear formula object for specifying the
fixed-effects in the marker-specific submodel. The response outcomes are separated by 
\code{+} on the left of \code{~} and the covariates are separated by \code{+} on 
the right of the \code{~}. For data-driven trajectory, polynomial time functions 
should be specified with \code{FP()} function, with powers specified between brackets. 
For sigmoid trajectory, no time function should be specified.}

\item{random}{a list containing one-sided formula for the random-effects in the
marker-specific submodel. Covariates with a random-effect are separated
by \code{+}. For sigmoid trajectory, NULL should be specified.}

\item{range}{a list containing vector indicating the range of the outcomes (that is
the minimum and maximum).}

\item{link}{optional vector of families of parameterized link functions defining
the measurement models (one by outcome). Option "id" (default) indicates no link function.
Other possibilities include "linear" specifies a linear link function, and "Splines" for 
approximating the link function by I-splines. 
For splines case, the number of nodes and the nodes location should be also 
specified. The number of nodes is first entered followed by \code{-}, then the location 
is specified with "equi", "quant" or "manual" for respectively equidistant
nodes, nodes at quantiles of the marker distribution or interior nodes
entered manually in argument \code{intnodes}. It is followed by \code{-splines}.
For example, "7-equi-splines" means
I-splines with 7 equidistant nodes, "6-quant-splines" means I-splines with 6
nodes located at the quantiles of the marker distribution and
"9-manual-splines" means I-splines with 9 nodes, the vector of 7 interior
nodes being entered in the argument \code{intnodes}.}

\item{intnodes}{optional vector of interior nodes. This argument is only
required for a I-splines link function with nodes entered manually.}

\item{model}{vector defining the model basis (one per outcome). Option "logistic" indicates
a sigmoid trajectory model, and "linear" indicates a data-driven trajectory model.}

\item{timevar}{name of the variable representing the measurement times.}

\item{fixed.s}{one-sided formula for covariates in the
temporal recalibration submodel. Covariates are separated
by \code{+}.}

\item{subject}{name of the covariate representing the grouping structure.}

\item{data}{data frame containing all variables named in other arguments.}

\item{subset}{optional vector giving the subset of observations in
\code{data} to use. By default, all lines.}

\item{na.action}{Integer indicating how NAs are managed. The default is 1
for 'na.omit'. The alternative is 2 for 'na.fail'. Other options such as
'na.pass' or 'na.exclude' are not implemented in the current version.}

\item{B}{optional specification for the initial values for the parameters.}

\item{posfix}{Optional vector giving the indices in vector B of the
parameters that should not be estimated. Default to NULL, all parameters are
estimated.}

\item{survival}{two-sided formula object. The left side of the formula corresponds 
to a Surv() object for right-censored (\code{Surv(EntryTime,Time,Indicator)})
and possibly left-truncated (\code{Surv(EntryTime,Time,Indicator)}).
Multiple causes of event can be considered 
in the Indicator (0 for censored, k for event of cause k). The right side of the 
formula specifies the covariates to include in the survival model.
Cause-specific covariate effect are specified with \code{cause()}.For example,  
Surv(Time,Indicator) ~ X1 + cause(X2) indicates a common effect of X1 and a cause-specific effect of X2.}

\item{hazard}{optional family of hazard function assumed for the survival model. 
By default, "Weibull" specifies a Weibull baseline risk function. Other possibilities 
are "piecewise" for a piecewise constant risk function or "splines" for a cubic M-splines 
baseline risk function. For these two latter families, the number of nodes and the 
location of the nodes should be specified as well, separated by -. The number of 
nodes is entered first followed by -, then the location is specified with "equi", 
"quant" or "manual" for respectively equidistant nodes, nodes at quantiles of the 
times of event distribution or interior nodes entered manually in argument hazardnodes. 
It is followed by - and finally "piecewise" or "splines" indicates the family of 
baseline risk function considered. Examples include "5-equi-splines" for M-splines 
with 5 equidistant nodes, "6-quant-piecewise" for piecewise constant risk over 5 
intervals and nodes defined at the quantiles of the times of events distribution 
and "9-manual-splines" for M-splines risk function with 9 nodes, the vector of 7 
interior nodes being entered in the argument hazardnodes. In the presence of competing 
events, a vector of hazards should be provided such as hazard=c("Weibull","5-quant-splines") 
with 2 causes of event, the first one modelled by a Weibull baseline cause-specific 
risk function and the second one by splines.}

\item{hazardrange}{optional vector indicating the range of the survival times 
(that is the minimum and maximum). By default, the range is defined according 
to the minimum and maximum observed values of the survival times. The option 
should be used only for piecewise constant and Splines hazard functions.}

\item{hazardnodes}{optional vector containing interior nodes if splines or piecewise 
is specified for the baseline hazard function in hazard.}

\item{logscale}{optional boolean indicating whether an exponential
(logscale=TRUE) or a square (logscale=FALSE -by default) transformation is
used to ensure positivity of parameters in the baseline risk functions. See
details section.}

\item{startWeibull}{optional numeric with Weibull hazard functions only.
Indicates the shift in the Weibull distribution.}

\item{surv.shift}{optional one-sided formula specifying the time function in interaction to the 
individual shift in the survival submodel. Default is NULL.}

\item{sharedtype}{vector indicating for each outcome and each event the shared random function type. 
\code{'RE'} indicates an association through the random effects included in the marker-specific submodel.
\code{'CV'} defines a association through the predicted current value of the marker, in the latent process scale.}

\item{nGK}{Number of integration points from Gauss-Kronrod procedure (for association through marker current value).
Possibilities are 7 or 15 (default).}

\item{methInteg}{character indicating the type of integration to compute the
log-likelihood. 'MCO' for ordinary Monte Carlo, 'MCA' for antithetic Monte Carlo,
'QMC' for quasi Monte Carlo. Default to "QMC".}

\item{nMC}{integer, number of Monte Carlo simulations. Default to 1000.}

\item{maxiter}{optional maximum number of iterations for the Marquardt
iterative algorithm. By default, maxiter=100.}

\item{convB}{optional threshold for the convergence criterion based on the
parameter stability. By default, convB=0.0001.}

\item{convL}{optional threshold for the convergence criterion based on the
log-likelihood stability. By default, convL=0.0001.}

\item{convG}{optional threshold for the convergence criterion based on the
derivatives. By default, convG=0.0001.}

\item{partialH}{optional vector giving the indices in vector B of parameters that
can be dropped from the Hessian matrix to define convergence criteria.}

\item{nsim}{number of points used to plot the estimated link functions. By
default, nsim=100.}

\item{verbose}{logical indicating if information about computation should be
reported. Default to TRUE.}

\item{returndata}{logical indicating if data used for computation should be
returned. Default to FALSE, data are not returned.}

\item{nproc}{number of cores for parallel computation.}

\item{clustertype}{the type of cluster that should internally be created.
See \code{parallel::makeCluster} for possible values.}
}
\value{
An object of class "jointLTSM".
}
\description{
This function fits extended joint disease progression models with shared random effects.
The aim of the joint disease progression model is to describe the markers’ trajectories
over the latent disease time, reconstituted by relocating patients along the temporal
axis according to patient’s disease advancement and marker observations, while
simultaneously accounting for the association with the risk of event occurence during
follow-up. The model consists of three submodels : one for the temporal recalibration,
one for the longitudinal markers, and one for the time-to-event process.
The individual temporal recalibration results from observed time regressed on covariates,
and an individual deviation.
The longitudinal submodel handles multiple continuous longitudinal outcomes
(Gaussian or non-Gaussian, curvilinear) in a mixed effects framework.
The survival submodel handles right-censored time-to-events with competing risks.
The association between the longitudinal and the survival data is captured by including 
the random effect from the mixed model or the predicted current level 
of the underlying process as a linear predictor in the proportional hazard survival model.
Parameters of the submodels are estimated simultaneously using a maximum likelihood method,
through a Marquardt-Levenberg algorithm.
}
\references{
Saulnier T., Fabbri M., Pavy-Le Traon A., Le Goff M., Helmer C., Péran P., Meissner W. G., 
Rascol O., Foubert-Samier A., Proust-Lima C. (2023). Disease Progression in Multiple System 
Atrophy : The Value of Clinical Cohorts with Long Follow-Up. Movement Disorders, 38(8), 1567-1569.
}
\author{
Tiphaine Saulnier, Cecile Proust-Lima
}
