% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTSM-package.R
\docType{package}
\name{LTSM-package}
\alias{LTSM-package}
\title{Estimation of latent time shift models}
\description{
Functions for the estimation of latent time shift models (LTSM).
Continuous outcomes are handled for the longitudinal part,
whereas the survival part considers multiple competing events.
The likelihood is computed using Monte-Carlo integration. Estimation is achieved
by maximizing the log-likelihood using a robust iterative algorithm.
}
\details{
Please report to the LTSM-team any question or suggestion regarding the package
via github only (https://github.com/TiphaineSaulnier/LTSM/issues).
}
\references{
Kühnel, Raket, Åström, Berger, Hansen, Krismer, ... & EMSA‐SG Natural History Study Investigators. (2022).
Disease Progression in Multiple System Atrophy—Novel Modeling Framework and Predictive Factors. Movement Disorders.

Philipps, Hejblum, Prague, Commenges, Proust-Lima (2021).
Robust and efficient optimization using a Marquardt-Levenberg algorithm with 
R package marqLevAlg, The R Journal 13:2.
}
\author{
Tiphaine Saulnier, Cecile Proust-Lima, Viviane Philipps
}
\keyword{package}
