% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTSM.R
\name{LTSM}
\alias{LTSM}
\title{Estimation of latent time shift model for multivariate longitudinal 
outcomes (and time-to-event data) with a time-shifting process.}
\usage{
LTSM(
  response,
  range,
  var.time,
  fixed_s,
  subject,
  data,
  subset = NULL,
  na.action = 1,
  B,
  posfix = NULL,
  survival = NULL,
  hazard = "Weibull",
  hazardrange = NULL,
  hazardnodes = NULL,
  logscale = FALSE,
  startWeibull = 0,
  methInteg = "QMC",
  nMC = 1000,
  maxiter = 100,
  convB = 1e-04,
  convL = 1e-04,
  convG = 1e-04,
  partialH = NULL,
  nsim = 100,
  verbose = TRUE,
  returndata = FALSE,
  nproc = 1,
  clustertype = NULL
)
}
\arguments{
\item{response}{character vector containing the names of the outcome-responses. ## Y}

\item{range}{vector indicating the range of the outcomes (that is
the minimum and maximum). ## Ak and Kk}

\item{var.time}{name of the time variable.   ## t}

\item{fixed_s}{a one-sided formula for specifying the fixed-effects defining 
the time-shift s.}

\item{subject}{name of the covariate representing the grouping structure.}

\item{data}{data frame containing all variables named in \code{response},
\code{fixed_s}, \code{survival} and \code{subject}.}

\item{subset}{optional vector giving the subset of observations in
\code{data} to use. By default, all lines.}

\item{na.action}{Integer indicating how NAs are managed. The default is 1
for 'na.omit'. The alternative is 2 for 'na.fail'. Other options such as
'na.pass' or 'na.exclude' are not implemented in the current version.}

\item{B}{optional specification for the initial values for the parameters.
Initial values should be entered in the order detailed in \code{details} section.}

\item{posfix}{Optional vector giving the indices in vector B of the
parameters that should not be estimated. Default to NULL, all parameters are
estimated.}

\item{survival}{two-sided formula object. The left side of the formula corresponds 
to a Surv() object for right-censored (\code{Surv(EntryTime,Time,Indicator)})
and possibly left-truncated (\code{Surv(EntryTime,Time,Indicator)}).
Multiple causes of event can be considered 
in the Indicator (0 for censored, k for event of cause k). The right side of the 
formula specifies the covariates to include in the survival model.
Cause-specific covariate effect are specified with \code{cause()}.For example,  
Surv(Time,Indicator) ~ X1 + cause(X2) indicates a common effect of X1 and a cause-specific effect of X2.}

\item{hazard}{optional family of hazard function assumed for the survival model. 
By default, "Weibull" specifies a Weibull baseline risk function. Other possibilities 
are "piecewise" for a piecewise constant risk function or "splines" for a cubic M-splines 
baseline risk function. For these two latter families, the number of nodes and the 
location of the nodes should be specified as well, separated by -. The number of 
nodes is entered first followed by -, then the location is specified with "equi", 
"quant" or "manual" for respectively equidistant nodes, nodes at quantiles of the 
times of event distribution or interior nodes entered manually in argument hazardnodes. 
It is followed by - and finally "piecewise" or "splines" indicates the family of 
baseline risk function considered. Examples include "5-equi-splines" for M-splines 
with 5 equidistant nodes, "6-quant-piecewise" for piecewise constant risk over 5 
intervals and nodes defined at the quantiles of the times of events distribution 
and "9-manual-splines" for M-splines risk function with 9 nodes, the vector of 7 
interior nodes being entered in the argument hazardnodes. In the presence of competing 
events, a vector of hazards should be provided such as hazard=c("Weibull","5-quant-splines") 
with 2 causes of event, the first one modelled by a Weibull baseline cause-specific 
risk function and the second one by splines.}

\item{hazardrange}{optional vector indicating the range of the survival times 
(that is the minimum and maximum). By default, the range is defined according 
to the minimum and maximum observed values of the survival times. The option 
should be used only for piecewise constant and Splines hazard functions.}

\item{hazardnodes}{optional vector containing interior nodes if splines or piecewise 
is specified for the baseline hazard function in hazard.}

\item{logscale}{optional boolean indicating whether an exponential
(logscale=TRUE) or a square (logscale=FALSE -by default) transformation is
used to ensure positivity of parameters in the baseline risk functions. See
details section}

\item{startWeibull}{optional numeric with Weibull hazard functions only.
Indicates the shift in the Weibull distribution.}

\item{methInteg}{character indicating the type of integration to compute the
log-likelihood. 'MCO' for ordinary Monte Carlo, 'MCA' for antithetic Monte Carlo,
'QMC' for quasi Monte Carlo. Default to "QMC".}

\item{nMC}{integer, number of Monte Carlo simulations. Default to 1000.}

\item{maxiter}{optional maximum number of iterations for the Marquardt
iterative algorithm. By default, maxiter=100.}

\item{convB}{optional threshold for the convergence criterion based on the
parameter stability. By default, convB=0.0001.}

\item{convL}{optional threshold for the convergence criterion based on the
log-likelihood stability. By default, convL=0.0001.}

\item{convG}{optional threshold for the convergence criterion based on the
derivatives. By default, convG=0.0001.}

\item{partialH}{optional vector giving the indices in vector B of parameters that
can be dropped from the Hessian matrix to define convergence criteria.}

\item{nsim}{number of points used to plot the estimated link functions. By
default, nsim=100.    ## ?????}

\item{verbose}{logical indicating if information about computation should be
reported. Default to TRUE.}

\item{returndata}{logical indicating if data used for computation should be
returned. Default to FALSE, data are not returned.}

\item{nproc}{number of cores for parallel computation.}

\item{clustertype}{the type of cluster that should internally be created.
See \code{parallel::makeCluster} for possible values.}
}
\value{
An object of class "LTSM" is returned containing some internal information
used in related functions. Users may investigate the following elements : 
\item{ns}{number of grouping units in the dataset} 
\item{loglik}{log-likelihood of the model} 
\item{best}{vector of parameter estimates in the same order as specified in 
\code{B} and detailed in section \code{details}}
\item{V}{vector containing the upper triangle matrix of variance-covariance
estimates of \code{best} with exception for variance-covariance parameters
of the random-effects for which \code{V} contains the variance-covariance
estimates of the Cholesky transformed parameters displayed in \code{cholesky}} 
\item{gconv}{vector of convergence criteria: 1. on the parameters, 2. on the 
likelihood, 3. on the derivatives} 
\item{conv}{status of convergence: =1 if the convergence criteria were satisfied, 
=2 if the maximum number of iterations was reached, =4 or 5 if a problem occured 
during optimisation} 
\item{call}{the matched call} 
\item{niter}{number of Marquardt iterations} 
\item{nevent}{number of occured event}
\item{pred}{table of individual predictions and residuals in the underlying
latent process scale; it includes marginal predictions (pred_m), marginal
residuals (resid_m), subject-specific predictions (pred_ss) and
subject-specific residuals (resid_ss) and finally the transformed
observations in the latent process scale (obs).}
\item{predRE_Y}{table containing individual predictions of the outcome-specific
random intercept}
\item{predSurv}{table containing the predicted baseline risk function and
the predicted cumulative baseline risk function }
\item{cholesky}{vector containing the estimates of the Cholesky transformed
parameters of the variance-covariance matrix of the random-effects}
\item{estimlink}{table containing the simulated values of each outcome and
the corresponding estimated link function} 
\item{AIC}{the Akaike's information criterion}
\item{BIC}{the Bayesian information criterion}
\item{CPUtime}{the runtime in seconds}
\item{data}{the original data set (if returndata is TRUE)}
}
\description{
This function fits XXX.
}
\details{
A. GENERALIZED LOGISTIC MODELS

XXX
multivariate version : one sub-model per outcome



B. THE SURVIVAL MODEL

association btw longitudinal and survival data : shared random effect (?)



C. THE VECTOR OF PARAMETERS B

The parameters in the vector of initial values \code{B} or in the vector of
maximum likelihood estimates \code{best} are included in the following
order:
(1) parameters for the baseline risk function: 2 parameters for each Weibull, 
nz-1 for each piecewise constant risk and nz+2 for each splines risk. In the 
presence of competing events, the number of parameters should be adapted to 
the number of causes of event;
(2) for all covariates in survival, one parameter 
is required. Covariates parameters should be included in the same order as in survival.
In the presence of cause-specific effects, the number of parameters should be
multiplied by the number of causes;
(3) parameter(s) of association between the longitudinal 
and the survival process: one parameter per random effect
and per cause of event is required;
(4) parameters for the time-shift s_i: one parameter per covariate in \code{fixed_s} as a fixed effect
(no intercept), and one variance parameter as the random intercept;
(5) the outcome-specific rate B_k: one parameter per response-outcome;
(6) the outcome-specific random location v_ik: one mean parameter per outcome, and one
variance parameter per outcome;
(7) the outcome-specific random intercepts u_ik: one variance parameter per outcome;
(8) the outcome-specific standard errors e_ijk: one variance parameter per outcome.
}
\examples{
#### Examples



}
\references{
Kühnel, Raket, Åström, Berger, Hansen, Krismer, ... & EMSA‐SG Natural History Study Investigators. (2022).
Disease Progression in Multiple System Atrophy—Novel Modeling Framework and Predictive Factors. Movement Disorders.

Philipps, Hejblum, Prague, Commenges, Proust-Lima (2021).
Robust and efficient optimization using a Marquardt-Levenberg algorithm with 
R package marqLevAlg, The R Journal 13:2.
}
\author{
Viviane Philipps, Tiphaine Saulnier and Cecile Proust-Lima
}
