% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRE.jointLTSM.R
\name{predictRE.LTSM}
\alias{predictRE.LTSM}
\title{Predictions of the random-effects}
\usage{
predictRE.LTSM(mod, newdata, subject = NULL, re.init = NULL, re.posfix = NULL)
}
\arguments{
\item{mod}{an object inheriting from class \code{jointLTSM}.}

\item{newdata}{data frame containing the data from which predictions are to be computed. 
The data frame should include at least all the covariates and 
the marker(s) values and the grouping structure. Names should match exactly the names 
of the variables in the model.}

\item{subject}{character specifying the name of the grouping structure.
If NULL (the default), the same as in the model will be used.}

\item{re.init}{optional vector of random effects, default is a random pick according to 
random-effects distribution}

\item{re.posfix}{optional vector giving the indices in vector re.init of the
random effects that should not be estimated. Default to NULL, all random effects are
estimated.}
}
\value{
a list containing:
\item{ns}{number of grouping units in the dataset}
\item{best}{vector of parameter estimates}
\item{predRE}{table containing individual predictions of the random-effects : 
a line per subject.}
\item{predRE_varcov}{list of vectors containing the upper triangle matrix of variance-covariance
estimates of the individually predicted random effects}
}
\description{
The function computes the predicted values of the random effects given observed data
provided in input.
}
\author{
Tiphaine Saulnier, Cecile Proust-Lima
}
