% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRE.LTSM.R
\name{predRE.LTSM}
\alias{predRE.LTSM}
\title{Predictions of the random-effects}
\usage{
predRE.LTSM(model, newdata, subject = NULL)
}
\arguments{
\item{model}{an object inheriting from class \code{LTSM}.}

\item{newdata}{data frame containing the data from which predictions are to be computed. 
The data frame should include at least all the covariates, 
the marker(s) values and the grouping structure. Names should match exactly the names 
of the variables in the model.}

\item{subject}{character specifying the name of the grouping structure.
If NULL (the default), the same as in the model will be used.}
}
\value{
a matrix containing the predicted random-effects.
}
\description{
The function computes the predicted values of the random effects given observed data
provided in input.
}
\author{
Tiphaine Saulnier, Cecile Proust-Lima, Viviane Philipps
}
