% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jointLTSM.R
\name{summary.jointLTSM}
\alias{summary.jointLTSM}
\title{Summary of a (joint) latent time shift model}
\usage{
\method{summary}{jointLTSM}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{jointLTSM} for a joint latent time shift model.}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
The function prints a summary of the model estimates.
}
\description{
This function provides a summary of model estimated with the \code{jointLTSM} function.
}
\author{
Tiphaine Saulnier, Cecile Proust-Lima
}
